add some new methods
in previous version, insert sorter are based on binary search, thus when
the sequence is highly sorted it won't get any improvement. this version
made a new class to deal with this.
shell sorter added, but it just don't reduce the number of action to be
made in best situation.